requirements.txt
# playwright
# playwright>=1.52.0 # should be already installed
# camoufox[geoip]>=0.4.11 # should be already installed | optional, for click-based solvers

# twocaptcha (my own fork with async support because 2captcha-python doesn't accept my PR)
2captcha-python-async==1.5.1

# files
aiofiles==24.1.0

# environment
python-dotenv==1.1.0

# tests
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-cov==6.1.1

